backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
Pillow==10.2.0
torch==2.2.2
torchvision==0.17.2 